requirements.txt
pydantic>=2.0
sqlalchemy>=2.0
playwright>=1.40
openai>=1.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
